commit
commit

+++ src/main/java/tn/esprit/spring/controller/AccountingController.java
@@ -58,7 +58,7 @@ public class AccountingController {
 
 	@PutMapping("/affecterAccountingToUser/{accounting_id}}/{id_user}")
 	@ResponseBody
-	public void affecterAccountingToUser(@PathVariable("accounting_id}") int idp,@PathVariable("id_user") int id) {
+	public void affecterAccountingToUser(@PathVariable("accounting_id}") int idp,@PathVariable("id_user") Long id) {
 		accountingService.affecterAccountingToUser(idp, id);
 	}
 	@GetMapping("/prime")
