+++ src/main/java/tn/esprit/spring/repository/OfferRepository.java
@@ -3,5 +3,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 import tn.esprit.spring.entity.Offer;
 @Repository
-public interface OfferRepository extends CrudRepository<Offer, Integer> {
+public interface OfferRepository extends CrudRepository<Offer, Long> {
 }
